synth_queue fn_8026CF78: widen secIndex param to int
Declaring secIndex as int (not u8) lets the caller pass the raw value to
synthSetStudioChannelScale while the SYNTH_SEQUENCE_QUEUE macro's & 0xFF
provides the index mask, matching the target's split (mr r30,r3 raw +
clrlwi for indexing) rather than reusing one masked copy. 3.4 -> 27.6%.
Residual is a body/condition CSE of cursor->time (one reload) that resists
source levers in an audio TU (no peephole/scheduling pragmas).

+++ src/main/audio/synth_queue.c
@@ -314,7 +314,7 @@ u32 seqStartPlay(u8* norm, u8* drum, u8* midiSetup, u8* song, SynthPlayPara* par
  *
  * EN v1.0 Address: 0x8026CF78, size 232b
  */
-void fn_8026CF78(u8 secIndex)
+void fn_8026CF78(int secIndex)
 {
     SynthSequenceQueue* section;
